refactor: extract action helpers and add tests
